Adding likes count and button and comments count to the scream dialog and refactoring the like button into its own component
